feat:타입 type폴더안에 삽입
