+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './clubResponse';
 export * from './profileResponse';
+export * from './loginResponse';
