+++ src/types/loginResponse.ts
@@ -1,0 +1,6 @@
+export type LoginResponse = {
+  accessToken: string;
+  expiresIn: number;
+  issuedAt: number;
+  refreshToken: string;
+};
